% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_specs_2016.R
\docType{data}
\name{core_specs_2016}
\alias{core_specs_2016}
\title{specification of core dataset 2016}
\format{
A data frame with 98 rows and 9 variables
}
\source{
\url{https://www.hcup-us.ahrq.gov/db/nation/nis/nisfilespecs.jsp#2016NIS}
}
\usage{
core_specs_2016
}
\description{
A dataset containing the specifications of 2016 NIS core dataset
It is used to load ASC file
}
\keyword{datasets}
