% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_specs.R
\docType{data}
\name{hospital_specs}
\alias{hospital_specs}
\title{specification of hospital dataset}
\format{
A data frame with 14 rows and 4 variables
}
\source{
\url{https://www.hcup-us.ahrq.gov/db/nation/nis/nisfilespecs.jsp#2016NIS}
}
\usage{
hospital_specs
}
\description{
A dataset containing the specifications of NIS hospital dataset
It is used to load ASC file
}
\keyword{datasets}
