% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_function.R
\name{load_nis}
\alias{load_nis}
\title{load nis function}
\usage{
load_nis(filepath, year, data_type)
}
\arguments{
\item{filepath}{character Path to the input file}

\item{year}{integer Year of data}

\item{data_type}{character type of data to load Eg:'hospital','core','severity}
}
\value{
A data frame
}
\description{
A function to load NIS data between 2016 and 2018
}
\examples{
load_nis('C:/data.asc',2016,'hospital')
}
