% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity_specs.R
\docType{data}
\name{severity_specs}
\alias{severity_specs}
\title{specification of severity dataset}
\format{
A data frame with 5 rows and 4 variables
}
\source{
\url{https://www.hcup-us.ahrq.gov/db/nation/nis/nisfilespecs.jsp#2016NIS}
}
\usage{
severity_specs
}
\description{
A dataset containing the specifications of NIS severity dataset
It is used to load ASC file
}
\keyword{datasets}
